+++ CHANGES.rst
@@ -20,6 +20,9 @@ Changelog
   Plone groups will only be created for selected organizations.  When selecting or
   unselecting organizations, linked Plone groups are created/deleted accordingly.
   [gbastien]
+- Moved setting.getOwnOrganizationPath to
+  utils.get_own_organization/utils.get_own_organization_path.
+  [gbastien]
 
 1.9 (2018-09-04)
 ----------------